separated GUI in different package + coverage report
As of now, all tests are covering 68% Instructions / 68% Branches / 70%
Lines / 77% Methods / 100% Types / 68% Complexity of entire metrics
package (9 Classes + 1 Enum).
