+++ src/test/java/metrics/AllTests.java
@@ -5,7 +5,7 @@ import org.junit.runner.RunWith;
 import org.junit.platform.suite.api.SelectClasses;
 
 @RunWith(JUnitPlatform.class)
-@SelectClasses({MetricsTest.class,IndicatorTest.class, MethodDataTest.class, RuleTest.class, ReadJavaProjectTest.class,FileDealerTest.class})
+@SelectClasses({MetricsTest.class,IndicatorTest.class, MethodDataTest.class, RuleTest.class, ReadJavaProjectTest.class,FileDealerTest.class,CompareFilesTest.class,QualityTest.class})
 public class AllTests {
 
 }
